requirements/development.txt
-r base.txt
django-debug-toolbar>=1.4
Werkzeug>=0.11.2
django-autofixture==0.12.1
